refactor: split main into modules
